Fixes failing tests after folder rename
