+++ src/app/__tests__/TopicDashboard/ResourceItem.test.js
@@ -1,21 +1,18 @@
 import React from 'react';
-import ReactDom from 'react-dom';
 import { shallow } from 'enzyme';
-import { ResourceItem } from '../../TopicDasboard/ResourceItem';
+import { ResourceItem } from '../../TopicDashboard/ResourceItem';
 import resourcesData from '../../../dummyDB/resourcesData';
 
 const FirstResource = resourcesData[0];
 
 const initialState = {
-  editing: false
+	editing: false
 };
 describe('ResourceItem', () => {
-  let wrapper;
-  beforeEach(() => {
-    wrapper = shallow(
-      <ResourceItem {...initialState} resource={FirstResource} />
-    );
-  });
+	let wrapper;
+	beforeEach(() => {
+		wrapper = shallow(<ResourceItem {...initialState} resource={FirstResource} />);
+	});
 
-  it('renders without crashing', () => {});
+	it('renders without crashing', () => {});
 });
